Odrobine uporzadkowalem kod, nastepnie zajme sie stylowaniem raportu pdf
