+++ resources/views/pages/accountMenu.blade.php
@@ -2,10 +2,14 @@
 
 @section('title', 'Zarządzanie kontem')
 
+{{--@section('content')--}}
+{{--    <div class="container__centered">--}}
+{{--        <form action="/changePassword" method="GET" class="container__button">--}}
+{{--            <button type="submit" class="container__button-btn">Zmień hasło</button>--}}
+{{--        </form>--}}
+{{--    </div>--}}
+{{--@endsection--}}
+
 @section('content')
-    <div class="container__centered">
-        <form action="/changePassword" method="GET" class="container__button">
-            <button type="submit" class="container__button-btn">Zmień hasło</button>
-        </form>
-    </div>
+    <a href="/podglad-raportu" target="_blank">🔍 Podgląd raportu</a>
 @endsection
